requirements.txt
pyyaml>=6.0
requests>=2.31.0
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.11.0
torch>=2.1.0
transformers>=4.44.0
datasets>=2.21.0
peft>=0.12.0
accelerate>=0.34.0
trl>=0.10.1
python-dotenv>=1.0.1
openai>=1.40.0
